add /chat create

+++ crypto/random/random.go
@@ -2,7 +2,7 @@ package random
 
 import "crypto/rand"
 
-func GenerateRandomBytes(length int) ([]byte, error) {
+func GenerateRandomBytes(length int64) ([]byte, error) {
 	b := make([]byte, length)
 	if _, err := rand.Read(b); err != nil {
 		return nil, ErrGeneratingFailed
